refactor: split homeboy cli helpers
